+++ gulp/metalsmith.js
@@ -10,6 +10,7 @@ const filter = require("gulp-filter");
 const assign = require("lodash.assign");
 const handlebars = require("handlebars");
 const handlebarsLayouts = require("handlebars-layouts");
+const markdownHelper = require("helper-markdown");
 
 const htmlMinifier = require("metalsmith-html-minifier");
 const markdown = require("metalsmith-markdown");
@@ -21,6 +22,7 @@ const debug = require("metalsmith-debug");
 const discoverPartials = require("metalsmith-discover-partials");
 
 const gravityAssets = require("./gravity-assets.js");
+const jobListings = require("./metalsmith-job-listings.js");
 
 function metalsmith() {
   // config for the site environment we're building
@@ -33,6 +35,7 @@ function metalsmith() {
   // register Handlebars helpers
   handlebars.registerHelper("moment", require("helper-moment"));
   handlebars.registerHelper(handlebarsLayouts(handlebars));
+  handlebars.registerHelper("markdown", markdownHelper);
 
   // register special partials
   gravityAssets.registerSvgSymbolsAsPartial(handlebars);
@@ -64,9 +67,10 @@ function metalsmith() {
               excludeRobots: siteEnv.excludeRobots
             }
           })
+          .use(jobListings())
           .use(pageTitles())
           .use(markdown())
-          .use(permalinks(":title"))
+          .use(permalinks(":page-url"))
           .use(
             discoverPartials({
               directory: `${paths.templates.src}${paths.templates.partials}`